requirements.txt
aiohttp==3.9.3
aiohttp-retry==2.8.3
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.0
async-timeout==4.0.3
attrs==23.2.0
Automat==24.8.1
backoff==2.2.1
bcrypt==4.2.1
beautifulsoup4==4.13.3
blinker==1.7.0
bs4==0.0.2
build==1.2.2.post1
cachetools==5.5.1
certifi==2024.2.2
cffi==1.17.1
charset-normalizer==3.3.2
chroma-hnswlib==0.7.6
chromadb==0.6.3
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
constantly==23.10.4
cryptography==44.0.1
cssselect==1.2.0
defusedxml==0.7.1
Deprecated==1.2.18
durationpy==0.9
exceptiongroup==1.2.2
fastapi==0.115.8
filelock==3.17.0
Flask==3.0.2
Flask-Cors==5.0.0
Flask-Limiter==3.10.1
Flask-Mail==0.10.0
Flask-SQLAlchemy==3.1.1
flatbuffers==25.2.10
frozenlist==1.4.1
fsspec==2025.2.0
google==3.0.0
google-auth==2.38.0
google-genai==1.1.0
googleapis-common-protos==1.67.0
greenlet==3.0.3
grpcio==1.70.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.28.1
humanfriendly==10.0
hyperlink==21.0.0
idna==3.6
importlib_metadata==8.5.0
importlib_resources==6.5.2
incremental==24.7.2
itemadapter==0.11.0
itemloaders==1.3.2
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==1.0.1
kubernetes==32.0.0
limits==4.0.1
lxml==5.3.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mmh3==5.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.5
numpy==2.2.2
oauthlib==3.2.2
onnxruntime==1.20.1
opentelemetry-api==1.30.0
opentelemetry-exporter-otlp-proto-common==1.30.0
opentelemetry-exporter-otlp-proto-grpc==1.30.0
opentelemetry-instrumentation==0.51b0
opentelemetry-instrumentation-asgi==0.51b0
opentelemetry-instrumentation-fastapi==0.51b0
opentelemetry-proto==1.30.0
opentelemetry-sdk==1.30.0
opentelemetry-semantic-conventions==0.51b0
opentelemetry-util-http==0.51b0
ordered-set==4.1.0
orjson==3.10.15
overrides==7.7.0
packaging==24.2
parsel==1.10.0
posthog==3.13.0
Protego==0.4.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyDispatcher==2.0.7
Pygments==2.19.1
PyJWT==2.8.0
pyOpenSSL==25.0.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
queuelib==1.7.0
requests==2.31.0
requests-file==2.1.0
requests-oauthlib==2.0.0
rich==13.9.4
rsa==4.9
Scrapy==2.12.0
service-identity==24.2.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.27
starlette==0.45.3
sympy==1.13.3
tenacity==9.0.0
tldextract==5.1.3
tokenizers==0.21.0
tomli==2.2.1
tqdm==4.67.1
twilio==9.0.0
Twisted==24.11.0
typer==0.15.1
typing_extensions==4.12.2
urllib3==2.2.1
uvicorn==0.34.0
w3lib==2.3.1
watchfiles==1.0.4
websocket-client==1.8.0
websockets==14.2
Werkzeug==3.0.1
wrapt==1.17.2
yarl==1.9.4
zipp==3.21.0
zope.interface==7.2
